finally delete redundant models.go
